+++ CppQueryableTest/Queryable/Functional/FirstFunctionalTest.cpp
@@ -21,15 +21,15 @@ using namespace QueryBuilder;
 class FirstFunctionalTest : public ::testing::Test
 {
 protected:
-  uint expectedUnorderedFirst = 9;
-  uint expectedOrderedFirst = 0;
-  uint expectedUnorderedOver40 = 76;
-  uint expectedOrderedOver40 = 45;
-  uint threshold = 40;
-  IQueryable<uint> queryable;
+  size_t expectedUnorderedFirst = 9;
+  size_t expectedOrderedFirst = 0;
+  size_t expectedUnorderedOver40 = 76;
+  size_t expectedOrderedOver40 = 45;
+  size_t threshold = 40;
+  IQueryable<size_t> queryable;
 
   FirstFunctionalTest() :
-    queryable(BuildQueryable2(std::vector<uint>({ expectedUnorderedFirst, 4, 7, 4, 3, 76, 8, 45, 34, 76, 0, 867 })))
+    queryable(BuildQueryable2(std::vector<size_t>({ expectedUnorderedFirst, 4, 7, 4, 3, 76, 8, 45, 34, 76, 0, 867 })))
   {
   }
 
@@ -56,31 +56,31 @@ TEST_F(FirstFunctionalTest, FirstVectorUninitialized)
 
 TEST_F(FirstFunctionalTest, FirstVector)
 {
-  uint value = this->queryable.First();
+  size_t value = this->queryable.First();
   ASSERT_EQ(this->expectedUnorderedFirst, value);
 }
 
 TEST_F(FirstFunctionalTest, FirstSet)
 {
-  uint value = ISortedQueryable<uint>(BuildQueryable2(this->queryable.ToSet())).First();
+  size_t value = ISortedQueryable<size_t>(BuildQueryable2(this->queryable.ToSet())).First();
   ASSERT_EQ(this->expectedOrderedFirst, value);
 }
 
 TEST_F(FirstFunctionalTest, FirstMultiSet)
 {
-  uint value = ISortedQueryable<uint>(BuildQueryable2(this->queryable.ToMultiSet())).First();
+  size_t value = ISortedQueryable<size_t>(BuildQueryable2(this->queryable.ToMultiSet())).First();
   ASSERT_EQ(this->expectedOrderedFirst, value);
 }
 
 TEST_F(FirstFunctionalTest, FirstDeque)
 {
-  uint value = IQueryable<uint>(BuildQueryable2(this->queryable.ToDeque())).First();
+  size_t value = IQueryable<size_t>(BuildQueryable2(this->queryable.ToDeque())).First();
   ASSERT_EQ(this->expectedUnorderedFirst, value);
 }
 
 TEST_F(FirstFunctionalTest, FirstList)
 {
-  uint value = IQueryable<uint>(BuildQueryable2(this->queryable.ToList())).First();
+  size_t value = IQueryable<size_t>(BuildQueryable2(this->queryable.ToList())).First();
   ASSERT_EQ(this->expectedUnorderedFirst, value);
 }
 
@@ -102,7 +102,7 @@ TEST_F(FirstFunctionalTest, FirstWhereConditionFails)
 {
   try
   {
-    this->queryable.First([](uint value) { return false; });
+    this->queryable.First([](size_t value) { return false; });
   }
   catch (std::runtime_error& ex)
   {
@@ -112,70 +112,70 @@ TEST_F(FirstFunctionalTest, FirstWhereConditionFails)
 
 TEST_F(FirstFunctionalTest, FirstWhereVector)
 {
-  uint value = this->queryable.First([&](uint value) { return value > this->threshold; });
+  size_t value = this->queryable.First([&](size_t value) { return value > this->threshold; });
   ASSERT_EQ(this->expectedUnorderedOver40, value);
 }
 
 TEST_F(FirstFunctionalTest, FirstWhereSet)
 {
-  uint value = ISortedQueryable<uint>(BuildQueryable2(this->queryable.ToSet()))
-    .First([&](uint value) { return value > this->threshold; });
+  size_t value = ISortedQueryable<size_t>(BuildQueryable2(this->queryable.ToSet()))
+    .First([&](size_t value) { return value > this->threshold; });
   ASSERT_EQ(this->expectedOrderedOver40, value);
 }
 
 TEST_F(FirstFunctionalTest, FirstWhereMultiSet)
 {
-  uint value = ISortedQueryable<uint>(BuildQueryable2(this->queryable.ToMultiSet()))
-    .First([&](uint value) { return value > this->threshold; });
+  size_t value = ISortedQueryable<size_t>(BuildQueryable2(this->queryable.ToMultiSet()))
+    .First([&](size_t value) { return value > this->threshold; });
   ASSERT_EQ(this->expectedOrderedOver40, value);
 }
 
 TEST_F(FirstFunctionalTest, FirstWhereDeque)
 {
-  uint value = IQueryable<uint>(BuildQueryable2(this->queryable.ToDeque()))
-    .First([&](uint value) { return value > this->threshold; });
+  size_t value = IQueryable<size_t>(BuildQueryable2(this->queryable.ToDeque()))
+    .First([&](size_t value) { return value > this->threshold; });
   ASSERT_EQ(this->expectedUnorderedOver40, value);
 }
 
 TEST_F(FirstFunctionalTest, FirstWhereList)
 {
-  uint value = IQueryable<uint>(BuildQueryable2(this->queryable.ToList()))
-    .First([&](uint value) { return value > this->threshold; });
+  size_t value = IQueryable<size_t>(BuildQueryable2(this->queryable.ToList()))
+    .First([&](size_t value) { return value > this->threshold; });
   ASSERT_EQ(this->expectedUnorderedOver40, value);
 }
 
 TEST_F(FirstFunctionalTest, FirstWhereEven)
 {
-  uint expected = 4;
-  uint value = this->queryable
-    .Where([](uint value) { return value % 2 == 0; })
+  size_t expected = 4;
+  size_t value = this->queryable
+    .Where([](size_t value) { return value % 2 == 0; })
     .First();
   ASSERT_EQ(expected, value);
 }
 
 TEST_F(FirstFunctionalTest, FirstWhereOdd)
 {
-  uint expected = expectedUnorderedFirst;
-  uint value = this->queryable
-    .Where([](uint value) { return value % 2 == 1; })
+  size_t expected = expectedUnorderedFirst;
+  size_t value = this->queryable
+    .Where([](size_t value) { return value % 2 == 1; })
     .First();
   ASSERT_EQ(expected, value);
 }
 
 TEST_F(FirstFunctionalTest, FirstWhereWhereEven)
 {
-  uint expected = 76;
-  uint value = this->queryable
-    .Where([](uint value) { return value % 2 == 0; })
-    .First([](uint value) { return value > 10; });
+  size_t expected = 76;
+  size_t value = this->queryable
+    .Where([](size_t value) { return value % 2 == 0; })
+    .First([](size_t value) { return value > 10; });
   ASSERT_EQ(expected, value);
 }
 
 TEST_F(FirstFunctionalTest, FirstWhereWhereOdd)
 {
-  uint expected = 45;
-  uint value = this->queryable
-    .Where([](uint value) { return value % 2 == 1; })
-    .First([](uint value) { return value > 10; });
+  size_t expected = 45;
+  size_t value = this->queryable
+    .Where([](size_t value) { return value % 2 == 1; })
+    .First([](size_t value) { return value > 10; });
   ASSERT_EQ(expected, value);
 }
